# Preload nodes in lookups