@slawlor 2022-12-30
Lints and rustfmt? Change looks good