@kevinlewi 2022-12-30
Question: should we do something similar like this for history proof generation?