@eozturk1 2022-12-30
> Question: should we do something similar like this for history proof generation?

I need to look into how we load the nodes but I think we can. I'll add it next week.

Another possibility is batch history proofs similar to batch lookups.